homework-6.3
