+++ src/entities/album/model/types.ts
@@ -1,0 +1,5 @@
+export type Album = {
+	userId: number;
+	id: number;
+	title: string;
+};
