+++ src/entities/comments/api/commentsApi.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import type { Comment } from "@/entities/[entity]/model/types";
+import type { Comment } from "@/entities/comments/model/types";
 
 export const commentsApi = createApi({
 	reducerPath: "commentsApi",
@@ -8,7 +8,7 @@ export const commentsApi = createApi({
 	endpoints: (builder) => ({
 		getCommentsByPostId: builder.query<Comment[], number>({
 			query: (postId) => `comments/?postId=${postId}`,
-			providesTags: ( postId) => [{ type: "Comment", postId }],
+			providesTags: (postId) => [{ type: "Comment", postId }],
 		}),
 	}),
 });
