+++ src/entities/comments/model/types.ts
@@ -1,0 +1,7 @@
+export type Comment = {
+	postId: number;
+	id: number;
+	name: string;
+	email: string;
+	body: string;
+};
