+++ src/entities/posts/api/postsApi.ts
@@ -1,5 +1,5 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import type { Post } from "@/entities/[entity]/model/types";
+import type { Post } from "@/entities/posts/model/types";
 
 export const postsApi = createApi({
 	reducerPath: "postsApi",
