+++ src/entities/posts/model/types.ts
@@ -1,0 +1,6 @@
+export type Post = {
+	userId: number;
+	id: number;
+	title: string;
+	body: string;
+};
