+++ src/entities/toDos/model/types.ts
@@ -1,0 +1,6 @@
+export type ToDo = {
+	userId: number;
+	id: number;
+	title: string;
+	completed: boolean;
+};
