+++ src/entities/users/model/types.ts
@@ -1,0 +1,23 @@
+export type User = {
+	id: number;
+	name: string;
+	username: string;
+	email: string;
+	address: {
+		street: string;
+		suite: string;
+		city: string;
+		zipcode: string;
+		geo: {
+			lat: number;
+			lng: number;
+		};
+	};
+	phone: string;
+	website: string;
+	company: {
+		name: string;
+		catchPhrase: string;
+		bs: string;
+	};
+};
