+++ src/pages/PostsFeedPage/PostsFeedPage.tsx
@@ -1,5 +1,5 @@
 import { PostList } from "@/widgets/PostList/PostList";
 
-export const PostsFeedPage = ({}: {}) => {
+export const PostsFeedPage = () => {
 	return <PostList />;
 };
