+++ src/shared/ui/ItemList/ItemList.module.scss
@@ -1,8 +1,7 @@
 .item_list {
 	display: flex;
 	flex-direction: column;
-	gap: 40px;
-	margin-top: 50px;
+	gap: 20px;
 	&__title {
 		display: flex;
 		justify-content: center;
