+++ src/widgets/PostList/PostList.tsx
@@ -2,7 +2,7 @@ import { PostCard } from "@/entities/post/ui/PostCard/PostCard";
 import { PostLengthFilter } from "@/features/PostLengthFilter/ui/PostLengthFilter";
 import styles from "./PostList.module.scss";
 import { usePosts } from "./model/hooks/usePosts";
-import type { Post } from "@/entities/[entity]/model/types";
+import type { Post } from "@/entities/posts/model/types";
 import { AsyncWrapper } from "@/widgets/AsyncWrapper/AsyncWrapper";
 import { ThemeWrapper } from "@/widgets/ThemeWrapper/ThemeWrapper";
