+++ src/widgets/UserSelection/UserSelection.tsx
@@ -3,7 +3,7 @@ import type { ChangeEvent } from "react";
 import { useUser } from "@/shared/lib/user/useUser";
 import styles from "./UserSelection.module.scss";
 import { useNavigate } from "react-router-dom";
-import { useGetUsersQuery } from "@/entities/[entity]/api/usersApi";
+import { useGetUsersQuery } from "@/entities/users/api/usersApi";
 
 export const UserSelection = () => {
 	const { setUser } = useUser();
@@ -12,7 +12,7 @@ export const UserSelection = () => {
 
 	useEffect(() => {
 		if(users) setUser(users[0]);
-	}, [users]);
+	}, [users, setUser]);
 
 	const handleChangeUser = (e: ChangeEvent<HTMLSelectElement>) => {
 		const currentUser = users?.find((user) => user.id === Number(e.target.value));
